HW2/ML: add new branch and edit mistake
