+++ run_app.py
@@ -1,0 +1,4 @@
+import uvicorn
+
+if __name__ == '__main__':
+    uvicorn.run("main:app", host="localhost", port=8000)
