+++ test/test_main.py
@@ -7,7 +7,7 @@ client = TestClient(app)
 def test_read_main():
     response = client.get("/")
     assert response.status_code == 200
-    assert response.json() == {"message": "World"}
+    assert response.json() == {"message": "Hello World"}
 
 
 def test_predict_positive():
